requirements.txt
Flask==2.0.2
boto3==1.18.57
nltk==3.6.3
pydub==0.25.1
python-dotenv==0.19.0
moviepy==1.0.3
pronouncing==0.2.0
